authentication and role based authorization are working fine

+++ Bookie/src/main/java/com/bookie/bookie/services/impl/HotelServiceImpl.java
@@ -5,7 +5,10 @@ import com.bookie.bookie.dtos.hotel.HotelDto;
 import com.bookie.bookie.dtos.hotel.HotelSearchDto;
 import com.bookie.bookie.entities.Hotel;
 import com.bookie.bookie.entities.Room;
+import com.bookie.bookie.entities.User;
+import com.bookie.bookie.entities.enums.Role;
 import com.bookie.bookie.exceptions.ResourceNotFoundException;
+import com.bookie.bookie.exceptions.UnauthorizedException;
 import com.bookie.bookie.mappers.HotelMapper;
 import com.bookie.bookie.repositories.HotelRepository;
 import com.bookie.bookie.repositories.RoomRepository;
@@ -30,19 +33,25 @@ public class HotelServiceImpl implements HotelService {
     private static final String HOTEL_NOT_FOUND_ERROR_MESSAGE = "Hotel not found with id : ";
     @Override
     @Transactional
-    public HotelDto createNewHotel(CreateHotelDto hotelDto) {
-        Hotel hotel = hotelRepository.save(hotelMapper.toEntity(hotelDto));
-        return hotelMapper.toDto(hotel);
+    public HotelDto createNewHotel(CreateHotelDto hotelDto, User authenticatedUser) {
+        Hotel hotel = hotelMapper.toEntity(hotelDto);
+        hotel.setActive(false);
+        hotel.setOwner(authenticatedUser);
+        Hotel savedHotel = hotelRepository.save(hotel);
+        return hotelMapper.toDto(savedHotel);
     }
 
     @Override
-    public HotelDto getHotelById(Long id) {
+    public HotelDto getHotelById(Long id, User authenticatedUser) {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(HOTEL_NOT_FOUND_ERROR_MESSAGE + id));
+        if (!authenticatedUser.equals(hotel.getOwner())){
+            throw new UnauthorizedException("the user with email : " + authenticatedUser.getEmail() + " is not the owner of the hotel " + hotel.getName());
+        }
         return hotelMapper.toDto(hotel);
     }
 
     @Override
-    public List<HotelDto> getAllHotels() {
+    public List<HotelDto> getAllHotels(User authenticatedUser) {
         List<Hotel> hotels = hotelRepository.findAll();
         return hotels.stream().map(hotelMapper::toDto).toList();
     }
@@ -55,23 +64,32 @@ public class HotelServiceImpl implements HotelService {
 
     @Override
     @Transactional
-    public HotelDto updateHotelById(Long id, CreateHotelDto hotelDto) {
+    public HotelDto updateHotelById(Long id, CreateHotelDto hotelDto, User authenticatedUser) {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(HOTEL_NOT_FOUND_ERROR_MESSAGE + id));
         hotelMapper.updateEntityFromDto(hotelDto, hotel);
+        if (!authenticatedUser.equals(hotel.getOwner())){
+            throw new UnauthorizedException("the user with email : " + authenticatedUser.getEmail() + " is not the owner of the hotel " + hotel.getName());
+        }
         return hotelMapper.toDto(hotel);
     }
 
     @Override
     @Transactional
-    public void deleteHotelById(Long id) {
+    public void deleteHotelById(Long id, User authenticatedUser) {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(HOTEL_NOT_FOUND_ERROR_MESSAGE + id));
+        if (!authenticatedUser.equals(hotel.getOwner())){
+            throw new UnauthorizedException("the user with email : " + authenticatedUser.getEmail() + " is not the owner of the hotel " + hotel.getName());
+        }
         hotel.setActive(false);
         roomRepository.deleteAll(hotel.getRooms());
     }
     @Override
     @Transactional
-    public HotelDto activateHotel(Long id){
+    public HotelDto activateHotel(Long id, User authenticatedUser){
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(HOTEL_NOT_FOUND_ERROR_MESSAGE + id));
+        if (!authenticatedUser.equals(hotel.getOwner())){
+            throw new UnauthorizedException("the user with email : " + authenticatedUser.getEmail() + " is not the owner of the hotel " + hotel.getName());
+        }
         hotel.setActive(true);
         for (Room room : hotel.getRooms()){
             inventoryService.initializeRoomForAYear(room);
